% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_cla.R
\docType{data}
\name{A_br_cla}
\alias{A_br_cla}
\title{example dataset}
\usage{
data(A_br_cla)
}
\description{
Dataset containing the classification of scars recorded on Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
