% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_crc.R
\docType{data}
\name{A_br_crc}
\alias{A_br_crc}
\title{example dataset}
\usage{
data(A_br_crc)
}
\description{
Dataset containing the information on the criss-cross and parallel scars recorderd on Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
