% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_pic.R
\docType{data}
\name{A_br_pic}
\alias{A_br_pic}
\title{example dataset}
\usage{
data(A_br_pic)
}
\description{
Picture of Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
