% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_sam.R
\docType{data}
\name{A_br_sam}
\alias{A_br_sam}
\title{example dataset}
\usage{
data(A_br_sam)
}
\description{
Coordinates of the scars recorderd on Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
