% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_sca.R
\docType{data}
\name{A_br_sca}
\alias{A_br_sca}
\title{example dataset}
\usage{
data(A_br_sca)
}
\description{
Scaled picture of Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
