% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_br_war.R
\docType{data}
\name{A_br_war}
\alias{A_br_war}
\title{example dataset}
\usage{
data(A_br_war)
}
\description{
Working area of the picture of Anoiapithecus brevirostris.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
