% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_cla.R
\docType{data}
\name{C_el_cla}
\alias{C_el_cla}
\title{example dataset}
\usage{
data(C_el_cla)
}
\description{
Dataset containing the classification of scars recorded on Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
