% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_crc.R
\docType{data}
\name{C_el_crc}
\alias{C_el_crc}
\title{example dataset}
\usage{
data(C_el_crc)
}
\description{
Dataset containing the information on the criss-cross and parallel scars recorderd on Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
