% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_pic.R
\docType{data}
\name{C_el_pic}
\alias{C_el_pic}
\title{example dataset}
\usage{
data(C_el_pic)
}
\description{
Picture of Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
