% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_sam.R
\docType{data}
\name{C_el_sam}
\alias{C_el_sam}
\title{example dataset}
\usage{
data(C_el_sam)
}
\description{
Coordinates of the scars recorderd on Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
