% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_sca.R
\docType{data}
\name{C_el_sca}
\alias{C_el_sca}
\title{example dataset}
\usage{
data(C_el_sca)
}
\description{
Scaled picture of Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
