% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_el_war.R
\docType{data}
\name{C_el_war}
\alias{C_el_war}
\title{example dataset}
\usage{
data(C_el_war)
}
\description{
Working area of the picture of Cervus elaphus eostephanoceros.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
\keyword{MicroWeaR}
