% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canc_plot.R
\name{canc_plot}
\alias{canc_plot}
\title{canc_plot}
\usage{
canc_plot(working.area, big_matrix, Type)
}
\arguments{
\item{working.area}{Ico.object: Ico class image}

\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the length; coordinates 3 and 4 for the width)}

\item{Type}{character: scars type}
}
\description{
Plot the scars stored in the object big_matrix associated to an image of class Ico. Now updated to use color coding to indicate classifications.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
