% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw.check.R
\name{mw.check}
\alias{mw.check}
\title{mw.check}
\usage{
mw.check(image.ico, big_matrix, type_traces, vector = NULL)
}
\arguments{
\item{image.ico}{Ico.object: class Ico image}

\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the lenght; coordinates 3 and 4 for the width)}

\item{type_traces}{character vector: classified scars types}

\item{vector}{numeric: which scars needs to be rechecked}
}
\value{
Ico.object: class Ico image

"priori" character vector: classified scars before running the check function

"posteriori" character vector: classified scars after running the check function
}
\description{
Check through an interactive multi-plot the classification provided by the autom_class function. Before runnin output.Ico with the a posteriori classification, remember to run again the cross.parallel function with the updated scars classifications.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
