% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.Ico.R
\name{output.Ico}
\alias{output.Ico}
\title{output.Ico}
\usage{
output.Ico(big_matrix, Type, cross_paral, w.area)
}
\arguments{
\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the lenght; coordinates 3 and 4 for the width)}

\item{Type}{character: a character vector}

\item{cross_paral}{matrix: output of the cross.parallel function}

\item{w.area}{output of Warea.Ico}
}
\value{
output_1 numeric: matrix with sampled marks count and classification

* N.pits is the total number of pits
* N.sp is the number of small pits
* N,lp is the number of large pits
* N.scratches is the total number of scratches
* N.fs is the number of fine scratches
* N.cs is the number of coarse scratches
* N.Ps is the parallel scratches
* N.Xs is the number of cross-parallel scratches
* %Ps is the percentages of scratches that are parallel
* %Xs is the percentage of scratches that are cross-parallel
* FeatureDensity is the total number of features per mm2
* Pmm2 is the number of pits per mm2
* Smm2 is the number of scratches per mm2
* PitPercent is the percentage of pits
* ScrachPercent is the percentage of scratches
}
\description{
Print a summary statistics table reporting the number of pits and scratches (and the size of any sub-category).
}
\examples{
\dontrun{
# A. brevirostris case-study
#load sampled scars
data("A_br_sam")
#load working area
data("A_br_war")
class<-autom_class(A_br_sam,A_br_war$image)
criss_cross<-cross.parallel(A_br_sam,A_br_war$image,class$Type)
output.Ico(A_br_sam,class$Type,criss_cross,A_br_war)
# C. elaphus eostephanoceros case-study
#load sampled scars
data("C_el_sam")
#load working area
data("C_el_war")
class<-autom_class(C_el_sam,C_el_war$image)
criss_cross<-cross.parallel(C_el_sam,C_el_war$image,class$Type)
output.Ico(C_el_sam,class$Type,criss_cross,C_el_war)
}
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel, Meaghan Wetherell
}
