% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parr_plot.R
\name{parr_plot}
\alias{parr_plot}
\title{parr_plot}
\usage{
parr_plot(image.ico, big_matrix, scr)
}
\arguments{
\item{image.ico}{Ico.object: Ico class}

\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the length; coordinates 3 and 4 for the width)}

\item{scr}{matrix: list of scratches returned from cross-parallel}
}
\description{
visualize pairs of scratches which are parallel and/or that cross each other.
}
\examples{
\dontrun{

}
}
\author{
Meaghan Wetherell
}
