% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ico.R
\name{plot_Ico}
\alias{plot_Ico}
\title{plot_Ico}
\usage{
plot_Ico(image.ico, xpos = 820, ypos = 0)
}
\arguments{
\item{image.ico}{Ico.object: Ico class image}

\item{xpos}{numeric: window x coordinates}

\item{ypos}{numeric: window xycoordinates}
}
\value{
Ico.object: plotted class Ico image
}
\description{
Plot Ico class image. Set the matrix containing the marks coordinates in the scars argument, the function returns them on the image.
}
\examples{
\dontrun{
# A. brevirostris case-study
#load image
data("A_br_pic")
#plot image
plot_Ico(A_br_pic)
# C. elaphus eostephanoceros case-study
#load image
data("C_el_pic")
#plot image
plot_Ico(C_el_pic)
}
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
