% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{rad2deg}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{numeric: value in radians}
}
\description{
Convert radians in degree.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
