% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.traces.R
\name{resume.traces}
\alias{resume.traces}
\title{resume.traces}
\usage{
resume.traces(
  image.ico,
  big_matrix,
  sounds = FALSE,
  delete.traces = c(1, 2, 3),
  delete.traces.int = FALSE,
  delay = 3
)
}
\arguments{
\item{image.ico}{Ico.object: class Ico image}

\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the lenght; coordinates 3 and 4 for the width)}

\item{sounds}{logical: if FALSE sounds are silences}

\item{delete.traces}{numeric vector: which scars needs to be deleted}

\item{delete.traces.int}{logical: set TRUE to delete traces by using an interactive way}

\item{delay}{numeric: specify seconds after that all devices will be closed}
}
\value{
big_matrix:  a matrix with stored coordinates
}
\description{
Utilities to export plots and to remove traces
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
