% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autom_class.R
\name{autom_class}
\alias{autom_class}
\title{autom_class}
\usage{
autom_class(big_matrix, image.ico, Pit_Scr = 4, Sm.Lg_Pit = 8,
  Fi.Co_Scr = 3, cexp = max(image.ico$res)/10)
}
\arguments{
\item{big_matrix}{matrix: a matrix with stored the coordinates (4) of the microwears}

\item{image.ico}{Ico.object: class Ico image}

\item{Pit_Scr}{numeric: parameter to discriminate between scratches and pits. If the ratio length/windth is minor or equal to the set value (4 for example) a scar is classified as pit; on the contrary as scratch}

\item{Sm.Lg_Pit}{numeric: parameter to classify a pit as small or large width based (in micron)}

\item{Fi.Co_Scr}{numeric: parameter to classify a scratch as course or fine width based (in micron)}

\item{cexp}{numeric: size of points}
}
\value{
Type character vector: types of the classified scars

Matrix matrix: length and width (in micron) of each scar
}
\description{
This function classifies into different categories the recorded marks using the samp.traces function.
}
\examples{
\dontrun{
# A. brevirostris case-study
#load sampled scars
data("A_br_sam")
#load working area
data("A_br_war")
class<-autom_class(A_br_sam,A_br_war$image)
# C. elaphus eostephanoceros case-study
#load sampled scars
data("C_el_sam")
#load working area
data("C_el_war")
class<-autom_class(C_el_sam,C_el_war$image)
}
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
