% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.parallel.R
\name{cross.parallel}
\alias{cross.parallel}
\title{cross.parallel}
\usage{
cross.parallel(big_matrix, working.area.ico, Type)
}
\arguments{
\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the length; coordinates 3 and 4 for the width)}

\item{working.area.ico}{Ico.object: Ico class working image, use $ to specify the image.}

\item{Type}{character: scars type}
}
\value{
numeric: matrix with number of pairs of parallel and/or scratches that cross each other
}
\description{
Detect pairs of scratches which are parallel and/or that cross each other.
}
\examples{
\dontrun{
# A. brevirostris case-study
#load sampled scars
data("A_br_sam")
#load working area
data("A_br_war")
class<-autom_class(A_br_sam,A_br_war$image)
criss_cross=cross.parallel(A_br_sam,A_br_war$image,class$Type)
criss_cross
# C. elaphus eostephanoceros case-study
#load sampled scars
data("C_el_sam")
#load working area
data("C_el_war")
class<-autom_class(C_el_sam,C_el_war$image)
criss_cross=cross.parallel(C_el_sam,C_el_war$image,class$Type)
criss_cross
}
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
