% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroWeaR.R
\docType{package}
\name{MicroWeaR-package}
\alias{MicroWeaR-package}
\alias{MicroWeaR}
\title{MicroWeaR is a Package to Examine and Score Microwear Scars}
\description{
MicroWeaR is a package to analyse dental microwears. The package contains functions to record and analyse microwear scars in a semi-automatic way.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
