% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroWeaR.R
\docType{package}
\name{MicroWeaR-package}
\alias{MicroWeaR-package}
\alias{MicroWeaR}
\title{MicroWeaR is a Package to Examine and Score Microwear Scars}
\description{
A tool that provides a summary statistic of the sampled scars for the user. The package includes functions to sample the scars, classify features as pits or scratches and then into consider also their respective subcategories, generate an output table with summary information, and obtain a visual surface-map in where scars are tracked.
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
