% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.Ico.R
\name{output.Ico}
\alias{output.Ico}
\title{output.Ico}
\usage{
output.Ico(big_matrix, Type, cross_paral, w.area)
}
\arguments{
\item{big_matrix}{matrix: a matrix with stored coordinates (4) of the sampled marks (coordinates 1 and 2 for the lenght; coordinates 3 and 4 for the width)}

\item{Type}{character: a character vector}

\item{cross_paral}{matrix: output of the cross.parallel function}

\item{w.area}{output of Warea.Ico}
}
\value{
output_1 numeric: matrix with sampled marks count and classification
}
\description{
Print a summary statistics table reporting the number of pits and scratches (and the size of any sub-category).
}
\examples{
\dontrun{
# A. brevirostris case-study
#load sampled scars
data("A_br_sam")
#load working area
data("A_br_war")
class<-autom_class(A_br_sam,A_br_war$image)
criss_cross<-cross.parallel(A_br_sam,A_br_war$image,class$Type)
output.Ico(A_br_sam,class$Type,criss_cross,A_br_war)
# C. elaphus eostephanoceros case-study
#load sampled scars
data("C_el_sam")
#load working area
data("C_el_war")
class<-autom_class(C_el_sam,C_el_war$image)
criss_cross<-cross.parallel(C_el_sam,C_el_war$image,class$Type)
output.Ico(C_el_sam,class$Type,criss_cross,C_el_war)
}
}
\author{
Antonio Profico, Flavia Strani, Pasquale Raia, Daniel DeMiguel
}
